backend/requirements.txt
# SIRA Platform Backend Requirements - Phase 2
# MVP Phase 2: Full Control Tower + AI + Telematics + Maritime

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2

# ASGI support
asgiref==3.7.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.9.0
supabase==2.7.4

# Validation
pydantic==2.7.0
pydantic-settings==2.4.0
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.0

# WebSocket
websockets==12.0

# AI Integrations
anthropic==0.34.0
openai==1.47.0
Pillow>=10.0.0

# IoT / Telematics
aiomqtt==2.0.0

# HTTP Client
httpx==0.27.0

# Geospatial
geojson==3.1.0
shapely==2.0.6

# Email
aiosmtplib==3.0.1

# PDF Generation
reportlab==4.0.8

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk[fastapi]==2.14.0
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9

# Redis (for WebSocket scaling & caching)
redis==5.0.8

# Task Queue (optional, for AI workers)
celery==5.4.0

# --- Phase 2: External Integrations ---

# AI Intelligence Engine
anthropic==0.42.0
openai==1.12.0

# MQTT (Flespi telematics)
aiomqtt==2.0.0

# Scheduling
apscheduler==3.10.4

# Testing
pytest==8.3.3
pytest-asyncio==0.23.8
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
ruff==0.6.0
